WIP: save a bit memory (in CPU caches)

+++ src/libknot/xdp/bpf-user.h
@@ -71,7 +71,7 @@ struct xsk_umem_info {
 
 	struct umem_frame *frames; /**< The memory frames. TODO: (uint8_t *frammem) might be more practical. */
 	uint32_t tx_free_count; /**< The number of free frames (for TX). */
-	uint32_t tx_free_indices[]; /**< Stack of indices of the free frames (for TX). */
+	uint16_t tx_free_indices[]; /**< Stack of indices of the free frames (for TX). */
 };
 
 typedef struct knot_xsk_socket {
